[refactoring] #190 Refactor splash ui code
