+++ app/src/main/java/com/cmc/curtaincall/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.cmc.curtaincall.ui
+package com.cmc.curtaincall
 
 import android.content.Context
 import android.os.Bundle
@@ -12,8 +12,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.cmc.curtaincall.AppNavHost
-import com.cmc.curtaincall.BuildConfig
 import com.cmc.curtaincall.common.design.theme.CurtainCallTheme
 import com.cmc.curtaincall.common.design.theme.NoRippleTheme
 import dagger.hilt.android.AndroidEntryPoint
