--- app/src/main/java/com/cmc/curtaincall/ui/splash/SplashSideEffect.kt
@@ -1,8 +1,0 @@
-package com.cmc.curtaincall.ui.splash
-
-import com.cmc.curtaincall.core.base.BaseSideEffect
-
-sealed class SplashSideEffect : BaseSideEffect {
-    object AutoLogin : SplashSideEffect()
-    object NeedLogin : SplashSideEffect()
-}
